+++ client/admin/customSounds/AdminSounds.js
@@ -2,7 +2,7 @@ import React, { useMemo, useCallback, useState, useEffect } from 'react';
 import { Box, Table, TextInput, Icon, Button } from '@rocket.chat/fuselage';
 
 import { useTranslation } from '../../contexts/TranslationContext';
-import { GenericTable, Th } from '../../../app/ui/client/components/GenericTable';
+import { GenericTable, Th } from '../../components/GenericTable';
 import { useCustomSound } from '../../contexts/CustomSoundContext';
 
 const FilterByText = ({ setFilter, ...props }) => {
